requirements.txt
social-auth-app-django
social_django